💩 Add code to debug the file mode of symbolic links to the node_modules directory created during testing
In Node.js with version less than 12, on Windows, an "EPERM: operation not permitted, stat '...\\node_modules'" error occurs when running a test.
This error occurs when running the “npm run” or “yarn run” command.
Perhaps there is a problem with the file permissions of the symbolic link to the node_modules directory.