🔧 Use npm v7 in this project
npm v7 is the latest version of npm that supports Node.js v10.
npm v8 and later no longer support Node.js 10.
Also, the latest npm overwrites `package-lock.json` files with forward incompatible `lockfileVersion: 3`.

So we specified the version of npm to be used for this project.
If the contributor is using Node.js 14.19 or later, the version of npm CLI should be automatically switched by Corepack.