✅ Avoid using RegExp in tests that check for the existing annotated tags
The output of the `git for-each-ref refs/tags` command contains an ever-changing SHA1 checksum.
For this reason, it was necessary to use regular expressions when checking for annotated tags that existed during testing.

However, we found that the `--format` option can be used to remove SHA1 from the output.
This makes it unnecessary to use regular expressions.