💚 Update "PATH" environment variable when executing `npm install --global ...` command on Windows
On Windows, the installed Corepack cannot be used without updating the `PATH` environment variable.
This is the same within the Bash shell.
In other words, the `corepack` command referenced in subsequent shell scripts is the Corepack before installation.
When the `corepack enable` command is executed, it only activates the old Corepack.

Therefore, the result of the `npm config get prefix` command must be added to the `PATH` environment variable.