✅ Specify a directory for testing only in the environment variable `COREPACK_HOME`
The `corepack prepare` command, for example, will install the package manager in the contributor's cache directory and change the "Last Known Good" release.
This is an irreversible change and the cache directory must be deleted to reset it.
The test code should not change any files in the contributor's computer.

So, use the environment variable `COREPACK_HOME` to change the Corepack cache directory.