✨ Ignore all errors from `npm config get ...` commands executed by the `tryNpmConfigGet()` function
The `pnpm version` command executes the `npm version` command as is.
Therefore, the configuration to be used should be obtained by the `npm config get ...` command.

`tryNpmConfigGet()` tries the `npm config get ...` command for this purpose.
However, if the `npm config get ...` command fails, we should not force termination there.
We should be allowed to use the `pnpm config get ...` command instead.