✅ Get the test name from the test context
From the Vitest test context, we can get the name of the test being executed.
https://vitest.dev/guide/test-context.html
This allows us to remove duplicate test name strings passed to the `tmpDir()` function.

Vitest's test context is not available from the `test.each()` function, so these test codes have been rewritten as `for...of` statements.